Edge thickness and related cleanups. Edge thickness now a static member of edge class that is set by CCModel draw (see GForge task for details.) Modifications to SizeSliderDialog for better edge functionality. Removal of unused edge thickness members in roadmap options class. Etc.
git-svn-id: svn+ssh://parasol-svn.cs.tamu.edu/research/parasol-svn/svnrepository/vizmo/trunk@407 8e255c04-9930-0410-9bf1-9572f6c5958d

+++ src/Models/Vizmo.h
@@ -73,7 +73,7 @@ class Vizmo {
     void ShowRoadMap(bool _show = true);
     bool IsRoadMapShown() const { return m_showMap; }
     void ChangeNodesSize(float _s, string _str);
-    void ChangeEdgeThickness(size_t _t);
+    void ChangeEdgeThickness(double _t);
     void ChangeNodesShape(string _s);
     bool IsRoadMapLoaded(){return m_mapModel;}
     void ChangeCCColor(double _r, double _g, double _b, string _s);
